pyproject.toml
[tool.poetry]
name = "scraping_service"
version = "0.1.0"
description = "Example template for standardized AI/ML services development"
authors = ["OBDeleven Data Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
undetected-chromedriver = "^3.5.5"
retry = "^0.9.2"
python-dotenv = "^1.0.1"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.linting.dependencies]
ruff = '^0.0.285'
isort = '^5.12.0'
black = '^23.7.0'


[tool.poetry.group.testing.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Read more: 'https://beta.ruff.rs/docs/rules/'
ignore = ['D104', 'D100', 'D103', 'G004', 'N999', 'I001', 'N806', 'COM812', 'ERA001', 'N803', 'D210', 'D211', 'D213']
select = [
    'B',   # flake8-bugbear
    'C4',  # flake8-comprehensions
    'ERA', # flake8-eradicate/eradicate
    'PIE', # flake8-pie
    'SIM', # flake8-simplify
    'TCH', # flake8-type-checking
    'TID', # flake8-tidy-imports
    'PTH', # flake8-use-pathlib
    'ARG', # flake8-unused-arguments
    'Q', # flake8-quotes
    'PT', # flake8-pytest-style
    'T20', # flake8-print
    'G', # flake8-logging-format
    'ICN', # flake8-import-conventions
    'COM', # flake8-commas
    'A', # flake8-builtins
    'D', # pydocstyle
    'I',   # isort
    'N',   # pep8-naming
    'RUF', # ruff checks
    'FLY', # flynt
    'PD', # pandas-vet
]

# Exclude a variety of commonly ignored directories.
exclude = [
    '.git',
    '.gitlab',
    '.mypy_cache',
    '.pytype',
    '.ruff_cache',
    '.venv',
    '__pypackages__',
    'venv',
]

line-length = 88 # Same as Black

[tool.ruff.flake8-quotes]
docstring-quotes = 'double'
inline-quotes = 'double'
multiline-quotes = 'double'

[tool.black]
line-length = 88
skip-string-normalization = true

[tool.isort] # Making isort compatible with black
profile = 'black'
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.pytest.ini_options]
pythonpath = [
  "."
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.interrogate]
ignore-init-module = true
exclude = [
    "src/tests",
]